Cecile Repellin -reinstate commented line in ParticleOnTorusNBodyHardCoreWithMagneticTranslationsHamiltonian that was meant for testing only
git-svn-id: https://www.nick-ux.org/diagham/svn/trunk@3021 b9410c57-6ef9-0310-98bc-a73ba792996a

+++ FQHE/src/Hamiltonian/ParticleOnTorusNBodyHardCoreWithMagneticTranslationsHamiltonian.cc
@@ -96,22 +96,22 @@ ParticleOnTorusNBodyHardCoreWithMagneticTranslationsHamiltonian::ParticleOnTorus
     }
   char* InteractionCoefficientFileName = new char [512];
   sprintf (InteractionCoefficientFileName, "%dbodydelta_interactioncoefficient_2s_%d_ratio_%.10f.dat", this->NBodyValue, this->NbrLzValue, this->Ratio);
-//   if (IsFile(InteractionCoefficientFileName))
-//     {
-//       ifstream File;
-//       File.open(InteractionCoefficientFileName, ios::binary | ios::in);
-//       if (!File.is_open())
-// 	{
-// 	  cout << "cannot open " << InteractionCoefficientFileName << endl;
-// 	}
-//       else
-// 	{
-// 	  for (int m1 = 0; m1 < this->NbrEntryPrecalculatedInteractionCoefficients1; ++m1)
-// 	    ReadBlockLittleEndian(File, this->PrecalculatedInteractionCoefficients[m1], this->NbrEntryPrecalculatedInteractionCoefficients2);
-// 	  File.close();
-// 	}
-//     }
-//   else
+  if (IsFile(InteractionCoefficientFileName))
+    {
+      ifstream File;
+      File.open(InteractionCoefficientFileName, ios::binary | ios::in);
+      if (!File.is_open())
+	{
+	  cout << "cannot open " << InteractionCoefficientFileName << endl;
+	}
+      else
+	{
+	  for (int m1 = 0; m1 < this->NbrEntryPrecalculatedInteractionCoefficients1; ++m1)
+	    ReadBlockLittleEndian(File, this->PrecalculatedInteractionCoefficients[m1], this->NbrEntryPrecalculatedInteractionCoefficients2);
+	  File.close();
+	}
+    }
+  else
     {
       ofstream File;
       File.open(InteractionCoefficientFileName, ios::binary | ios::out);
